migrate RequestLogger and ResponseLogger into ServletLogger
